Implemented emotional replies methodology
Agent: implemented Perceive method with a custom deplay.
           Modified Perform to have new EmotionalReply in mind.

SingleActionAgent: passed some of the generation of events to the action creator agent so that it has more control.

SingleReplyAgent: implemented react method as a reactive response to anticipation event.

Emotion: corrected bug where it's emotional stage was not being updated.

Added new Confidence Emotion as an example.

Event: added Event Type to define the message type (ACTION or REPLY)

Modified Time functions to have a more abstract control depending on the platform.
Time class is now a holder of a TimeCalculator class that holds the functionalities for time calculation. One TimeCalculator is the ChronoTimeCalculator that uses the c++ chrono namespace.

ActionProgress now can hold other ActionProgresses as it's stage's replies.

Disabled Qt opacity animation because of concurrent ActionProgress classes wanting to paint on the same object, which created an error.