Memory Leak fixed and bug correction
Added a vector of available emotions to the agent class to control the emotions correctly.

corrected event (pain in the a**) basically it was leaking memory because nowhere was it deleting it's pointer to the reply event if it had one.

increased World Manager abstraction to agents where it uses the new IAgent interface to simplify agent creation.