Added new emotions, actions and scenario agents
Added coop scenario sequence 1, which required the addition of new agents, actions and emotions.

World Manager now has a description that allows the explanation of what is and what is in the world.

Minor layout changes, where new emotions are added directly to the action instead of being one of it's children and the visual description of the world.