Emotion and events updated and big bug fix
Emotion now has a reference to the agent it is replying to (if it's a emotional reply)

ReplyEvent has been removed since Emotion now contains the necessary information to determine if it is a reply.

Corrected big bug where the agents weren't creating copies of the actions/emotions/stages they were using to express themselfs.