Updated Agent and WorldManager
Made variables public and removed functions.

Removed Agent children, no longer needed.

I decided to change this piece of to from a Framework to a Library, which implies a more open code, addionally I decide to change the way I program to be more flexible and that sometimes goes against traditional OO programming.